pyproject.toml
[project]
name = "real-estate-api"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "tenolly", email = "np50022005@yandex.ru"},
]
dependencies = [
    "fastapi==0.115.4",
    "pydantic-settings==2.6.1",
    "databases==0.9.0",
    "asyncpg==0.30.0",
    "uvicorn==0.32.0",
    "alembic>=1.14.0",
    "python-dotenv>=1.0.1",
    "pre-commit>=4.0.1",
    "pyyaml>=6.0.2",
    "psycopg2>=2.9.10",
    "pytest>=8.3.3",
    "httpx>=0.27.2",
    "pytest-asyncio>=0.24.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
